# backend/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediSync API",
    "description": "Hospital Management System API for managing doctors, patients, and appointments.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/appointments/{appointmentId}/no-show": {
      "put": {
        "tags": [
          "Appointments"
        ],
        "summary": "Mark appointment as No-Show",
        "description": "Marks an appointment as NO_SHOW if the patient did not attend. Requires DOCTOR or ADMIN role.",
        "operationId": "markNoShow",
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment marked as no-show",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires DOCTOR or ADMIN role"
          },
          "404": {
            "description": "Appointment not found"
          },
          "409": {
            "description": "Conflict - Appointment status cannot be changed (e.g., already completed)"
          }
        }
      }
    },
    "/api/appointments/{appointmentId}/cancel": {
      "put": {
        "tags": [
          "Appointments"
        ],
        "summary": "Cancel an appointment",
        "description": "Cancels a scheduled appointment. Access is restricted to the owning Patient, the assigned Doctor, or an Admin.",
        "operationId": "cancelAppointment",
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment successfully cancelled",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Not authorized to cancel this appointment"
          },
          "404": {
            "description": "Appointment not found"
          },
          "409": {
            "description": "Conflict - Appointment is already completed or cancelled"
          }
        }
      }
    },
...
    "/api/appointments": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Book a new appointment",
        "description": "Allows a Patient to schedule an appointment with a doctor. Requires PATIENT role.",
        "operationId": "bookAppointment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentBookDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Appointment successfully booked",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (e.g., Doctor not found, invalid time slot)"
          },
          "403": {
            "description": "Forbidden - Only patients can book appointments"
          }
        }
      }
    },
    "/api/appointments/{appointmentId}/complete": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Complete an appointment",
        "description": "Marks an appointment as COMPLETED and generates the associated medical record. Requires DOCTOR role.",
        "operationId": "completeAppointment",
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalRecordCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Appointment completed and record created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MedicalRecordDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Only the assigned doctor can complete this appointment"
          },
          "404": {
            "description": "Appointment not found"
          },
          "409": {
            "description": "Conflict - Appointment is already cancelled or completed"
          }
        }
      }
    },
...
    "/api/appointments/{appointmentId}": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "Get appointment details",
        "description": "Retrieves details of a specific appointment. Access is restricted to the owning Patient, the assigned Doctor, or an Admin.",
        "operationId": "getAppointmentById",
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved appointment details",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User is not associated with this appointment"
          },
          "404": {
            "description": "Appointment not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ]
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "county": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MedicalRecordCreateDto": {
        "type": "object",
        "properties": {
          "diagnosis": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "treatmentPlan": {
            "type": "string"
          },
          "prescription": {
            "type": "string"
          }
        },
        "required": [
          "diagnosis"
        ]
      },
      "MedicalRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "diagnosis": {
            "type": "string"
          },
          "treatmentPlan": {
            "type": "string"
          },
          "prescription": {
            "type": "string"
          },
          "appointmentId": {
            "type": "integer",
            "format": "int64"
          },
          "appointmentTime": {
            "type": "string",
            "format": "date-time"
          },
          "doctorId": {
            "type": "integer",
            "format": "int64"
          },
          "doctorName": {
            "type": "string"
          },
          "patientId": {
            "type": "integer",
            "format": "int64"
          },
          "patientName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DoctorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "specialization": {
            "type": "string"
          },
          "specializationLabel": {
            "type": "string"
          },
          "departmentId": {
            "type": "integer",
            "format": "int64"
          },
          "departmentName": {
            "type": "string"
          },
          "appointmentDuration": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AppointmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "appointmentTime": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELLED",
              "NO_SHOW"
            ]
          },
          "doctor": {
            "$ref": "#/components/schemas/DoctorDto"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "medicalRecord": {
            "$ref": "#/components/schemas/MedicalRecordDto"
          }
        }
      },
...
      "AppointmentBookDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "integer",
            "format": "int64"
          },
          "doctorId": {
            "type": "integer",
            "format": "int64"
          },
          "appointmentTime": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "appointmentTime",
          "doctorId",
          "patientId",
          "reason"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}